refactor: move S3StagingConfig to CDK
