+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/spec/S3StagingConfiguration.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.destination.redshift2.config
+package io.airbyte.cdk.load.spec
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonProperty
