+++ airbyte-integrations/connectors/destination-redshift-v2/gradle.properties
@@ -1,3 +1,3 @@
-cdkVersion=1.0.7
+cdkVersion=local
 testExecutionConcurrency=4
 JunitMethodExecutionTimeout=15m
